Enhance mover handling
