+++ Content.Shared/Movement/Systems/SharedMoverController.Input.cs
@@ -197,7 +197,7 @@ namespace Content.Shared.Movement.Systems
             Dirty(uid, mover);
         }
 
-        private bool TryUpdateRelative(EntityUid uid, InputMoverComponent mover, TransformComponent xform)
+        protected bool TryUpdateRelative(EntityUid uid, InputMoverComponent mover, TransformComponent xform)
         {
             // Starlight Start
             if (RelayQuery.TryComp(uid, out var relay)
